requirements.txt
requests==2.21.0
Pillow==6.0.0
instabot==0.62.0
python-dotenv==0.10.3